Ravine fissue doesn't destroy blocks under player

+++ src/main/java/com/crowsofwar/avatar/common/entity/EntityRavine.java
@@ -207,7 +207,7 @@ public class EntityRavine extends AvatarEntity {
 		
 		if (!worldObj.isRemote && breakBlocks) {
 			BlockPos last = new BlockPos(prevPosX, prevPosY, prevPosZ);
-			if (!last.equals(getPosition())) {
+			if (!last.equals(getPosition()) && !last.equals(initialPosition.toBlockPos())) {
 				
 				worldObj.destroyBlock(last.down(), true);
